Make file/path test work on multiple platforms
The -logfile option to runtests takes a path name, and it's supposed to
fail if the path is impossible to open.  Unfortunately,
"C:\dir__does__not__exist\foobar.txt" can open just fine on Unix, it's
just a file with the name "C:\dir__does__not__exist\foobar.txt".  So
changed that to "/dir__does__not__exist/foobar.txt".