Verify more extensively whether a 'name' is actually a path in `TestSuite.fromName`
As noted in #14, cases exist where `isdir` will return true where it is
not necessarily expected to do so. This is most evident when providing a
partial path that happens to be on the MATLAB path, where a relative
directory was implied. Whenever this is the case a
'MATLAB:cd:NonExistentDirectory' error is thrown. By adding additional
logic for verifying the different cases this can be circumvented.

Fixes #14