feat(buildFunctionHandleTestSuite): provide function-based test suites on R2016b and later
The `initTestSuite` can no longer be used in R2016b and later, due to
the way scripts are scoped in those versions. Instead the
`buildFunctionHandleTestSuite` function should be used.

In fact, this is the preferred way for generating these types of test
suites across all MATLAB versions that support them. `initTestSuite` has
been updated to generate warnings about the changed behavior in R2016b
and errors out with a descriptive error on affected versions. And
documentation is available in `buildFunctionHandleTestSuite` describing
how to support older MATLAB versions.

Fixes #19.